Fix tests
The integration test cannot use any Spring infrastructure and can only
interact with the app via HTTP.
Furthermore the actuator was added in order to give the application a
health endpoint which is needed by Arquillian in order to ensure that
the application is functioning before launching the test
